requirements.txt
mathplotlib==2.2.2
numpy==1.14.2
pillow==5.1.0
scipy==1.0.1
